fix(guardian-r1): resolve 8 design-system + file-size violations
