+++ src/renderer/features/personal/fitness/components/WorkoutEditDialog.tsx
@@ -4,8 +4,6 @@
 
 import { useState } from 'react';
 
-import { Plus, Trash2 } from 'lucide-react';
-
 import type { Exercise, ExerciseSet, Workout, WorkoutType } from '@shared/types';
 
 import {
@@ -27,6 +25,8 @@ import {
 
 import { useUpdateWorkout } from '../api/useFitness';
 
+import { WorkoutExerciseList } from './WorkoutExerciseList';
+
 // ── Helpers ─────────────────────────────────────────────────
 
 let nextKey = 0;
@@ -35,7 +35,7 @@ function uid(): string {
   return `ek-${String(nextKey)}`;
 }
 
-interface FormExercise extends Exercise {
+export interface FormExercise extends Exercise {
   _key: string;
   _setKeys: string[];
 }
@@ -205,36 +205,14 @@ export function WorkoutEditDialog({ workout, open, onOpenChange }: WorkoutEditDi
           </div>
 
           {/* Exercises */}
-          <div>
-            <div className="mb-2 flex items-center justify-between">
-              <span className="text-muted-foreground text-xs font-medium">Exercises</span>
-              <Button
-                className="text-primary h-auto p-0 text-xs font-medium"
-                size="sm"
-                type="button"
-                variant="ghost"
-                onClick={handleAddExercise}
-              >
-                <Plus className="h-3 w-3" />
-                Add Exercise
-              </Button>
-            </div>
-            <div className="space-y-3">
-              {exercises.map((exercise, exerciseIndex) => (
-                <EditExerciseInput
-                  key={exercise._key}
-                  exercise={exercise}
-                  exerciseIndex={exerciseIndex}
-                  onAddSet={() => handleAddSet(exerciseIndex)}
-                  onNameChange={(name) => handleExerciseNameChange(exerciseIndex, name)}
-                  onRemove={() => handleRemoveExercise(exerciseIndex)}
-                  onSetChange={(setIndex, field, value) =>
-                    handleSetChange(exerciseIndex, setIndex, field, value)
-                  }
-                />
-              ))}
-            </div>
-          </div>
+          <WorkoutExerciseList
+            exercises={exercises}
+            onAddExercise={handleAddExercise}
+            onAddSet={handleAddSet}
+            onExerciseNameChange={handleExerciseNameChange}
+            onRemoveExercise={handleRemoveExercise}
+            onSetChange={handleSetChange}
+          />
 
           {/* Notes */}
           <div>
@@ -276,83 +254,3 @@ export function WorkoutEditDialog({ workout, open, onOpenChange }: WorkoutEditDi
   );
 }
 
-// ── EditExerciseInput ─────────────────────────────────────────
-
-interface EditExerciseInputProps {
-  exercise: FormExercise;
-  exerciseIndex: number;
-  onNameChange: (name: string) => void;
-  onRemove: () => void;
-  onAddSet: () => void;
-  onSetChange: (setIndex: number, field: keyof ExerciseSet, value: string) => void;
-}
-
-function EditExerciseInput({
-  exercise,
-  exerciseIndex,
-  onNameChange,
-  onRemove,
-  onAddSet,
-  onSetChange,
-}: EditExerciseInputProps) {
-  return (
-    <div className="bg-muted/50 rounded-md p-3">
-      <div className="mb-2 flex items-center gap-2">
-        <Input
-          aria-label={`Exercise ${String(exerciseIndex + 1)} name`}
-          className="flex-1"
-          placeholder="Exercise name"
-          size="sm"
-          type="text"
-          value={exercise.name}
-          onChange={(e) => onNameChange(e.target.value)}
-        />
-        <Button
-          aria-label="Remove exercise"
-          className="text-muted-foreground hover:text-destructive h-auto p-1"
-          size="icon"
-          type="button"
-          variant="ghost"
-          onClick={onRemove}
-        >
-          <Trash2 className="h-3 w-3" />
-        </Button>
-      </div>
-      <div className="space-y-1">
-        {exercise.sets.map((exerciseSet, setIndex) => (
-          <div key={exercise._setKeys[setIndex]} className="flex items-center gap-2">
-            <span className="text-muted-foreground w-8 text-xs">S{String(setIndex + 1)}</span>
-            <Input
-              aria-label={`Set ${String(setIndex + 1)} reps`}
-              className="w-16"
-              placeholder="Reps"
-              size="sm"
-              type="number"
-              value={exerciseSet.reps ?? ''}
-              onChange={(e) => onSetChange(setIndex, 'reps', e.target.value)}
-            />
-            <Input
-              aria-label={`Set ${String(setIndex + 1)} weight`}
-              className="w-20"
-              placeholder="Weight"
-              size="sm"
-              type="number"
-              value={exerciseSet.weight ?? ''}
-              onChange={(e) => onSetChange(setIndex, 'weight', e.target.value)}
-            />
-            <span className="text-muted-foreground text-xs">lbs</span>
-          </div>
-        ))}
-      </div>
-      <Button
-        className="text-primary mt-1 h-auto p-0 text-xs font-medium"
-        size="sm"
-        type="button"
-        variant="ghost"
-        onClick={onAddSet}
-      >
-        + Add Set
-      </Button>
-    </div>
-  );
-}
